refactor: separate scrap and modify occupation to part
