+++ src/main/java/sejong/alom/teammate/domain/member/entity/ProfilePart.java
@@ -15,25 +15,25 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sejong.alom.teammate.domain.meta.entity.Occupation;
+import sejong.alom.teammate.domain.meta.entity.Part;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Table(name = "profile_occupation")
-public class ProfileOccupation {
+@Table(name = "profile_part")
+public class ProfilePart {
 	@Id
 	@GeneratedValue(
 		strategy = GenerationType.SEQUENCE,
-		generator = "profile_occupation_seq_gen"
+		generator = "profile_part_seq_gen"
 	)
 	@SequenceGenerator(
-		name = "profile_occupation_seq_gen",
-		sequenceName = "profile_occupation_seq"
+		name = "profile_part_seq_gen",
+		sequenceName = "profile_part_seq"
 	)
-	@Column(name = "profile_occupation_id")
+	@Column(name = "profile_part_id")
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
@@ -41,6 +41,6 @@ public class ProfileOccupation {
 	private Profile profile;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "occupation_id")
-	private Occupation occupation;
+	@JoinColumn(name = "part_id")
+	private Part part;
 }
