+++ src/main/java/sejong/alom/teammate/domain/scrap/entity/ProfileScrap.java
@@ -1,4 +1,4 @@
-package sejong.alom.teammate.domain.team.entity;
+package sejong.alom.teammate.domain.scrap.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -16,6 +16,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import sejong.alom.teammate.domain.member.entity.Member;
+import sejong.alom.teammate.domain.member.entity.Profile;
 import sejong.alom.teammate.global.util.BaseTimeEntity;
 
 @Entity
@@ -23,23 +24,23 @@ import sejong.alom.teammate.global.util.BaseTimeEntity;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Table(name = "team_scrap")
-public class TeamScrap extends BaseTimeEntity {
+@Table(name = "profile_scrap")
+public class ProfileScrap extends BaseTimeEntity {
 	@Id
 	@GeneratedValue(
 		strategy = GenerationType.SEQUENCE,
-		generator = "team_scrap_seq_gen"
+		generator = "profile_scrap_seq_gen"
 	)
 	@SequenceGenerator(
-		name = "team_scrap_seq_gen",
-		sequenceName = "team_scrap_seq"
+		name = "profile_scrap_seq_gen",
+		sequenceName = "profile_scrap_seq"
 	)
-	@Column(name = "team_scrap_id")
+	@Column(name = "profile_scrap_id")
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "team_id")
-	private Team team;
+	@JoinColumn(name = "profile_id")
+	private Profile profile;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "member_id")
